Few notes: - 'static' should always be the first keyword. Some compilers ignore this order, other stop with an error. -  It's a good habbit to add '= {0}' always when declaring arrays regardless if they are being set later or not. This forces the compiler to initialize the table with zeros. In C and C++ pointers and arrays are not requirerd to be initialized by zeros (unless appropriate compiler option is used).
Function 'fill_val' must be declared before it is used. That's a major difference between C++ and Fortran visibility of functions. If one wants to use a function that is in other file or below in the code, one can write the header of the function above the place it is used (this is called "forward declaration").

Additional note: adding keyword "extern" to such forward declarattion additionally tells compiler to skip checks if this function exists during compilation of the program as it is assumed the function may be loaded from external library during runtime.