gcc 4.8.1 fix
NOTE:
Technically speaking 'i = i++;' has undefined behaviour
(according to C99 standard). However, this piece of code:

  int i=0, j=0;

  i = i++;
  j = ++j;

  std::cout<<i<<" "<<j<<std::endl; 

for gcc 4.7.3 and earlier, always returned "1 1".
This is no longer true for gcc 4.8.1. It returns "0 1".
This is a potential problem for all our projects as thye might not be
gcc 4.8.1 compatible!

NOTE2: this is not true in 'for' loops. i++ is still valid expression.
       it's the assignment i = i++ that is invalid.